Add hipchat-data.txt
The raw text output from Jira when a new issue is created. This is what
we'll be listening for in our irssi plugin and silencing & filtering.

I probably missed some empty lines commented by Jira. Hard to keep track
when all the lines are the same and I have to scroll back in irssi with
pretty much no context.