Fix test `message public` listener
Turns out I got the argument list wrong. The argument list I was using
was for a `command_bind` (based on the tutorial mentioned in
0370866db92e1916af5ee98acf948adc3916d545).

I wanted to make a signal listener for public messages. Looked up the
documentation and figured out the correct arguments.

Also fixed my log message printing to append to the file instead of
overwriting it for every message received (previously I only got the
last message posted in my log output).