Checking for empty messages from HipChat (WIP)
Some trial and error in selecting the blank empty messages coming from
HipChat/JIRA. Had to do the output-to-file thing again, this time a new
different one. Was trying to match against an empty string message but
it turns out the empty messages that HipChat sends have a single space
in them, aggh!