hipchat-data.txt: Use generic issue prefix and number
Rename the issue prefix to SC for "some company" and change the issue
number to the one I used in the tag body. Forgot to change this when I
was removing real company-specific details. Probably don't need to
retroactively remove this from the repository as it's not a huge mistake
and going back to do so would be kind of a pain.